pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cronbot-cli"
version = "0.1.0"
description = "Internship diary automation CLI"
requires-python = ">=3.11"
dependencies = [
  "google-genai>=1.64.0",
  "playwright>=1.58.0",
  "python-dotenv>=1.2.1",
  "rich>=14.3.3",
  "typer>=0.24.1",
]

[project.scripts]
cronbot = "cronbot_entry:run"
cronjob = "cronbot_entry:run"

[tool.setuptools]
package-dir = {"" = "src"}
py-modules = ["cronbot_entry", "main"]

[tool.setuptools.packages.find]
where = ["src"]
